Create README.md
A book recommendation system that is a type of recommendation system where we get recommendations of similar books to the reader based on their interest.
It also includes a list of Top 50 books along with their user rating and up votes.
This project is built using Python, Flask, Pycharm and Jupyter Notebook. The web pages are developed using HTML and CSS.
The entire project was developed in Visual Studio Code.